:truck: Rename layout.js to Layout.js
